Return success status from openSubdiv_finishEvaluatorDescr()
Before this change the given function used to fail silently in
cases topology is bad or uniform evaluator initialization failed.
This used to leave evaluator in a state which is not usable for
further processing but callee function would never know this and
will likely crash later when evaluating subdivision limit surface.

Ideally error message or code need to be passed to the calle, but
that's marked as TODO for now.