Harden Far::SubdivisionTablesFactory to handle unconnected vertices
Unconnected vertices are non-manifold topology and trigger asserts in the
subdivision table factory. The code has been hardened to not assert out and
return the proper subdivision depth. The fix appears to be workin on our test
cases, but the odds are fairly high that non-manifolds will still break
something else...