WIN32 Glew and DLL fixes to CMakeLists.txt
Previously, dynamic targets were disabled on Windows due to configuration
errors in the CMake install targets. In addition, glew was forced to
static linkage on windows, even when dynamic Glew libraries were linked
which would result in linker errors.

In this change, static glew libraries are detected by looking for the "s"
suffix convention on the library (though, maybe this should be handled in
the FindGlew module).

In addition, there is now a fork in the opensubdiv osd_dynamic_{cpu|gpu}
install targets to account for the Windows DLL CMake quirks.