Fixed example dependencies with CUDA on Linux
Added an explicit target link dependency for OpenGL::GLX when
building example programs with CUDA on systems with GLX since
GLX is used by the example code used to initialize the CUDA
device context.

This is a simple workaround and we'd like to revisit and improve
how target dependencies are handled overall.

Fixes #1294