Fixed inconsistent warning levels for MSVC builds
The MSVC build has been setting /W3 as an override for /Wall but this
has been causing generator specfic build inconsistencies.

For examples, this override works when using the Visual Studio generators,
but not when using the Ninja generator. This causes Ninja to build using
/Wall instead of /W3.

This removes the initial setting of /Wall to avoid this inconsistency.

We plan to continue to address compiler warnings to allow us to raise
this warning level further but that is a separate task.