Remove libgomp.so link when building for TBB only.
When building with `NO_TBB=OFF` and `NO_OMP=ON` the `osd_cpu_obj` objects target
will be configured to link to `gomp`.

When OpenMP is disabled, the omp-specific files (`ompEvaluator` and `ompKernel`)
are not actually built, and therefore the link is not needed.

Although this doesn't create problems in opensubdiv builds, downstream projects
may face challenges as the link carries over through cmake to any project that
links to opensubdiv.

Signed-off-by: Niall Lenihan <nlenihan@wetafx.co.nz>